requirements.txt
langchain
chromadb
pypdf
sentence-transformers
google-generativeai